feat: Refactor TelegramLoggerService import paths and add new logger implementation
