+++ src/modules/common/common.module.ts
@@ -9,7 +9,7 @@ import { CountryModule } from '../country/country.module';
 import { CityModule } from '../city/city.module';
 import { MembershipModule } from '../membership/membership.module';
 import { CommonService } from './common.service';
-import { TelegramLoggerService } from 'src/utils/telegram-logger.service';
+import { TelegramLoggerService } from 'src/utils/telegram-logger';
 import { BroadcastModule } from '../broadcast/broadcast.module';
 import { UserModule } from '../user/user.module';
 import { AccessModule } from '../access/access.module';
